codalab/codalab/settings/base.py: parse bool from environment variables.
The value of an environment variable is always a string and any
non-empty string evals as true.  If we want to support false values in
the text we need to parse it ourselves.

This commit fixes boolean options in the django settings file: DEBUG,
IS_DEV, EMAIL_USE_TLS, BROKER_USE_SSL, and PIN_PASSCODE_ENABLED.  In
addition, AWS_S3_SECURE_URLS and AWS_QUERYSTRING_AUTH parsed the
string value but in different manner.  This commit uses the same
approach in all, for the sake of consistency.

We drop the use of distutils.util.strtobool because distutils is
deprecated and will be removed in Python 3.12 but our implementation
is compatible.