docker/flower/Dockerfile: parse bool from BROKER_USE_SSL env.
Environment variables are strings and any non-empty string evals as
true.  This commit only checks if the value is "True" and something
fancier check True or False or error would be nicer but it's messy to
do it on a dockerfile RUN command that is sh appending to a python
file.